apps/api/pyproject.toml
[project]
name = "ai-assistant-api"
version = "0.1.0"
description = "AI Assistant FastAPI service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "langgraph>=0.4.1",
  "mcp>=1.26.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
  "slowapi>=0.1.9",
  "sqlalchemy>=2.0.36",
  "tavily-python>=0.5.0",
  "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
